Add migrations, seeder

+++ EventPad/Data/EventPad.Context.Entities/CashoutEventReceipt.cs
@@ -11,7 +11,7 @@ public class CashoutEventReceipt : BaseEntity
     public virtual User User { get; set; }
 
     public string BankAccount { get; set; }
-    public decimal Amount { get; set; }
+    public float Amount { get; set; }
     public DateTime DateTime {  get; set; }
     public string RKTransactionId { get; set; }
 }
